requirements.txt
asgiref==3.4.1
autopep8==1.5.7
cachetools==4.2.4
certifi==2021.5.30
cffi==1.15.1
charset-normalizer==2.0.6
cryptography==3.4.8
defusedxml==0.7.1
diff-match-patch==20200713
dj-database-url==0.5.0
Django==3.2.8
django-allauth==0.45.0
django-ckeditor==6.1.0
django-cors-headers==3.10.0
django-countries==7.2.1
django-crispy-forms==1.13.0
django-debug-toolbar==3.2.2
django-extensions==3.1.3
django-import-export==2.6.1
django-js-asset==1.2.2
django-simple-history
django-storages==1.11.1
django-taggit==1.5.1
djangorestframework==3.12.4
docutils==0.17.1
drf-yasg==1.21.7
et-xmlfile==1.1.0
executing==0.8.2
google-api-core==2.1.0
google-auth==2.2.1
google-cloud-core==2.1.0
google-cloud-storage==1.42.3
google-crc32c==1.3.0
google-resumable-media==2.0.3
googleapis-common-protos==1.53.0
gunicorn==20.1.0
idna==3.2
inflection==0.5.1
Markdown==3.3.4
MarkupPy==1.14
numpy==1.21.2
oauthlib==3.1.1
odfpy==1.4.1
openpyxl==3.0.9
packaging==23.1
pandas==1.3.3
Pillow==8.3.2
protobuf==3.18.1
psycopg2-binary==2.9.1
pyasn1==0.4.8
pyasn1-modules==0.2.8
Pycco==0.6.0
pycodestyle==2.7.0
pycparser==2.20
Pygments==2.10.0
pygraphviz==1.7
PyJWT==2.1.0
pystache==0.6.0
python-dateutil==2.8.2
python3-openid==3.2.0
pytz==2021.3
PyYAML==5.4.1
requests==2.26.0
requests-oauthlib==1.3.0
rsa==4.7.2
six==1.16.0
smartypants==2.0.1
sqlparse==0.4.2
tablib==3.0.0
toml==0.10.2
uritemplate==4.1.1
urllib3==1.26.7
xlrd==2.0.1
xlwt==1.3.0
